requirements.txt
boto==2.41.0
Django==1.9.6
django-filter==0.13.0
django-notifications-hq
django-registration==2.1.2
django-shortuuidfield==0.1.3
django-storages==1.4.1
django-webpack-loader==0.3.0
djangorestframework==3.3.3
Markdown==2.6.6
shortuuid==0.4.3
six==1.10.0
django-postgrespool==0.3.0
psycopg2==2.5.1
gunicorn==19.3.0
